Minimal changes were required for this backport to work on 1.17.10...
Although this does not mean it is necessarily in completely perfect condition. My testing with this build was very limited but I did not encounter crashes and I did not encounter any compilation/runtime warnings related to use of the 1.7.10 spigot binary. In-addition, I did not find that the plugin was non-stable. It seems to be working but it is based on a backported version of the 1.16.1 codebase which is not, in itself, totally finished. If anyone is interested in working on this backport, they should let me know and I can add them as a contributor.